Adds ngx-translate to app

+++ apps/exam-generator/src/app/task/task-list/task-list.component.ts
@@ -4,6 +4,7 @@ import { Router } from "@angular/router";
 import { ExamTask } from "../models/exam-task.model";
 import { TaskService } from "../task.service";
 import { v4 as uuidv4 } from "uuid";
+import { ConfirmationService } from "primeng/api";
 
 @Component({
   selector: "app-task-list",
@@ -12,14 +13,14 @@ import { v4 as uuidv4 } from "uuid";
 export class TaskListComponent implements OnInit {
   public taskList: ExamTask[] = [];
 
-  public constructor(public router: Router, public taskService: TaskService) {}
+  public constructor(
+    public router: Router,
+    public taskService: TaskService,
+    private confirmationService: ConfirmationService
+  ) {}
 
   public async ngOnInit(): Promise<void> {
     await this.loadTasks();
-  }
-
-  private async loadTasks() {
-    this.taskList = await this.taskService.findAll();
   }
 
   public async createNewTask(): Promise<void> {
@@ -34,8 +35,23 @@ export class TaskListComponent implements OnInit {
     await this.router.navigate([`tasks/${task.id}/edit`]);
   }
 
-  public async deleteTask(task: ExamTask) {
-    await this.taskService.delete(task.id);
-    await this.loadTasks();
+  public async deleteTask(event: Event, task: ExamTask) {
+    this.confirmationService.confirm({
+      target: event.target || undefined,
+      message:
+        "Are you sure you want to delete this task? Removed items can't be restored.",
+      icon: "pi pi-exclamation-triangle",
+      accept: async () => {
+        await this.taskService.delete(task.id);
+        await this.loadTasks();
+      },
+      reject: () => {
+        //reject action
+      },
+    });
+  }
+
+  private async loadTasks() {
+    this.taskList = await this.taskService.findAll();
   }
 }
